The link to hide observers or locations now uses a lowercase 'x'
The circled 'x' was intuitive and an easy target to click on but it was
way too noisy visually when the page was displayed. That actually made
the page more difficult to read. A lowercase 'x' in normal size font
is much less distracting and still easy to use - though a little harder
to hit with the mouse.